try to implement the teaching algorithm.

+++ WorkstationController.Visual/BasewareUIElement.cs
@@ -211,8 +211,8 @@ namespace WorkstationController.VisualElement
                 carrier = (Carrier)_ware;
             int grid = carrier.GridID;
             int pinPos = (grid - 1) * Worktable.DistanceBetweenAdjacentPins + (int)_worktable.TopLeftPinPosition.X;
-            int xPos = pinPos;
-            int yPos = (int)_worktable.TopLeftPinPosition.Y;
+            double xPos = pinPos;
+            double yPos = _worktable.TopLeftPinPosition.Y;
             xPos = pinPos - (carrier.XOffset);  //get carrier x start pos
             yPos -= carrier.YOffset;
